Rename Admin to AbstractAdmin
This follows Symfony coding standard rules.

Ref: http://symfony.com/doc/current/contributing/code/standards.html#naming-conventions
