+++ Resources/doc/getting_started/the_form_view.rst
@@ -17,11 +17,11 @@ The basic class definition will look the same as the ``CategoryAdmin``:
     // src/AppBundle/Admin/BlogPostAdmin.php
     namespace AppBundle\Admin;
 
-    use Sonata\AdminBundle\Admin\Admin;
+    use Sonata\AdminBundle\Admin\AbstractAdmin;
     use Sonata\AdminBundle\Datagrid\ListMapper;
     use Sonata\AdminBundle\Form\FormMapper;
 
-    class BlogPostAdmin extends Admin
+    class BlogPostAdmin extends AbstractAdmin
     {
         protected function configureFormFields(FormMapper $formMapper)
         {
