+++ Resources/doc/reference/routing.rst
@@ -232,7 +232,7 @@ We now need to create our Controller, the easiest way is to extend the basic Son
 Removing a route
 ----------------
 
-Extending ``Sonata\AdminBundle\Admin\Admin`` will give your Admin classes the following
+Extending ``Sonata\AdminBundle\Admin\AbstractAdmin`` will give your Admin classes the following
 default routes:
 
 * batch
