+++ Resources/skeleton/Admin.php.twig
@@ -8,13 +8,13 @@
 
 namespace {{ namespace }};
 
-use Sonata\AdminBundle\Admin\Admin;
+use Sonata\AdminBundle\Admin\AbstractAdmin;
 use Sonata\AdminBundle\Datagrid\DatagridMapper;
 use Sonata\AdminBundle\Datagrid\ListMapper;
 use Sonata\AdminBundle\Form\FormMapper;
 use Sonata\AdminBundle\Show\ShowMapper;
 
-class {{ classBasename }} extends Admin
+class {{ classBasename }} extends AbstractAdmin
 {
     /**
      * @param DatagridMapper $datagridMapper
