+++ Tests/Admin/Extension/LockExtensionTest.php
@@ -53,7 +53,7 @@ class LockExtensionTest extends \PHPUnit_Framework_TestCase
 
         $this->modelManager = $this->getMock('Sonata\AdminBundle\Model\LockInterface');
 
-        $this->admin = $this->getMockBuilder('Sonata\AdminBundle\Admin\Admin')
+        $this->admin = $this->getMockBuilder('Sonata\AdminBundle\Admin\AbstractAdmin')
             ->disableOriginalConstructor()
             ->getMock();
