+++ Tests/Fixtures/Admin/CommentAdmin.php
@@ -2,10 +2,10 @@
 
 namespace Sonata\AdminBundle\Tests\Fixtures\Admin;
 
-use Sonata\AdminBundle\Admin\Admin;
+use Sonata\AdminBundle\Admin\AbstractAdmin;
 use Sonata\AdminBundle\Route\RouteCollection;
 
-class CommentAdmin extends Admin
+class CommentAdmin extends AbstractAdmin
 {
     public function setClassnameLabel($label)
     {
