+++ Tests/Fixtures/Admin/CommentWithCustomRouteAdmin.php
@@ -2,7 +2,7 @@
 
 namespace Sonata\AdminBundle\Tests\Fixtures\Admin;
 
-use Sonata\AdminBundle\Admin\Admin;
+use Sonata\AdminBundle\Admin\AbstractAdmin;
 use Sonata\AdminBundle\Route\RouteCollection;
 
 class CommentWithCustomRouteAdmin extends CommentAdmin
