+++ Tests/Fixtures/Admin/ModelAdmin.php
@@ -2,13 +2,13 @@
 
 namespace Sonata\AdminBundle\Tests\Fixtures\Admin;
 
-use Sonata\AdminBundle\Admin\Admin;
+use Sonata\AdminBundle\Admin\AbstractAdmin;
 use Sonata\AdminBundle\Datagrid\DatagridMapper;
 use Sonata\AdminBundle\Datagrid\ListMapper;
 use Sonata\AdminBundle\Form\FormMapper;
 use Sonata\AdminBundle\Show\ShowMapper;
 
-class ModelAdmin extends Admin
+class ModelAdmin extends AbstractAdmin
 {
     /**
      * @param DatagridMapper $datagridMapper
