+++ Tests/Fixtures/Admin/PostAdmin.php
@@ -2,9 +2,9 @@
 
 namespace Sonata\AdminBundle\Tests\Fixtures\Admin;
 
-use Sonata\AdminBundle\Admin\Admin;
+use Sonata\AdminBundle\Admin\AbstractAdmin;
 
-class PostAdmin extends Admin
+class PostAdmin extends AbstractAdmin
 {
     protected $metadataClass = null;
