+++ Tests/Show/ShowMapperTest.php
@@ -11,7 +11,6 @@
 
 namespace Sonata\AdminBundle\Tests\Show;
 
-use Sonata\AdminBundle\Admin\Admin;
 use Sonata\AdminBundle\Admin\AdminInterface;
 use Sonata\AdminBundle\Admin\FieldDescriptionCollection;
 use Sonata\AdminBundle\Builder\ShowBuilderInterface;
