+++ UPGRADE-3.1.md
@@ -1,0 +1,6 @@
+UPGRADE FROM 3.0 to 3.1
+=======================
+
+## Deprecated Admin class
+
+The `Admin` class is deprecated. Use `AbstractAdmin` instead.
